pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "llmchat"
version = "0.1.0"
description = "A minimal multi-provider LLM chat CLI with persistent sessions."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "you"}]
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "prompt_toolkit>=3.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
google = ["google-genai>=0.3"]
all = [
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=0.3",
    "httpx>=0.27",
]

[project.scripts]
llmchat = "llmchat.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["llmchat*"]
